requirements.txt
Django~=2.2.4
pytils~=0.3
django-tagging~=0.5.0
django-solo~=1.1.3
Pillow~=7.0.0
yattag~=1.13.2